modules/modeling/basic/linear_regression/requirements.txt
pyDataCanvas >= 0.4.0,<0.5.0
numpy
matplotlib
scikit-learn
